4045: Rename map constant file
