+++ shared/index.ts
@@ -17,7 +17,7 @@ export * from './utils/normalizeString'
 export { default as findNormalizedMatches, findAllMatches } from './utils/findNormalizedMatches'
 export * from './utils'
 export { getCategoryTiles } from './utils/categories'
-export * from './constants/maps'
+export * from './constants/map'
 export * from './constants/news'
 export * from './utils/dateFilterUtils'
 export * from './utils/getExcerpt'
